requirements.txt
selenium~=4.4.3
selenium-requests
webdriver-manager
pillow~=9.3.0
2captcha-python
